requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
opencv-python==4.8.1.78
torch==2.1.2
torchvision==0.16.2
faiss-cpu==1.7.4
sentence-transformers==2.2.2
python-multipart==0.0.6
numpy==1.26.3
pillow==10.2.0
python-dotenv==1.0.0